Move BluetoothCardReaderMessages out of internal package
